% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen2_processing.R
\name{calcIndex}
\alias{calcIndex}
\title{Spectral index and zonal statistics}
\usage{
calcIndex(
  safes,
  aoi,
  epsg = NULL,
  dx = NULL,
  dy = NULL,
  dt = NULL,
  aggregation = "mean",
  resampling = "bilinear",
  tmpagg = FALSE,
  index = NULL,
  label = "",
  outdir = ".",
  timeframe = "complete",
  years = NULL,
  threads = 1
)
}
\arguments{
\item{safes}{A character vector with the \code{.jp2} and \code{.tif} files of the
selected \code{SAFE} directories. Obtained with \code{list.files(<safedir>, "jp2$|.tif$", recursive =T, full.names = T)}}

\item{aoi}{A sf-object with POLYGONS or MULTIPOLYGONS. For objects with ZM
dimensions these dimensions are dropped before calculation.}

\item{epsg}{A character specifying the epsg code for the the cube view specified
as \code{"EPSG:<code>"}. The resulting \code{GTiff} files and the \code{GPKG}
with the zonal statistics will be projected into this CRS.
\code{dx} and \code{dy} are understood in the units of the CRS specified here.
If left empty, the CRS of the \code{aoi} variable will be used.}

\item{dx}{A numeric specifying the resolution of the output raster in x dimension
expressed in the map unit specified in \code{epsg.} If \code{epsg} is left
empty it defaults to the map unit of \code{aoi.}}

\item{dy}{A numeric specifying the resolution of the output raster in y dimension
expressed in the map unit specified in \code{epsg}.
If \code{epsg} is left empty it defaults to the map unit of \code{aoi}.}

\item{dt}{Resolution of the time dimension expressed as ISO8601 string
(e.g. \code{"P16D"} for 16 days).}

\item{aggregation}{A character vector specifying the aggregation method for
images of the same time stamp. Must be one of \code{"min"}, \code{"max"},
\code{"mean"}, \code{"median"}, or \code{"first"}. Defaults to \code{"mean"}.}

\item{resampling}{A character specifying the resampling method used to transform
images from their native CRS to the output CRS. Can be any of the methods
supported by gdalwarp (see \url{www.gdal.org/programs/gdalwarp.html}).
Defaults to \code{"bilinear"}.}

\item{tmpagg}{A logical indicating if temporal aggregation of all timestemps should
be performed calling \code{\link[gdalcubes:reduce_time]{reduce_time()}}. Defaults to \code{FALSE}.}

\item{index}{A character specifying the short name of the index to be calculated.
Available indices can be checked with \code{check_indices()}.}

\item{label}{A character label which is appended to the output files in the
form of \code{<label>_<index>_<timestep>}. Files with the same name in
\code{outdir} will be overwritten without warning.}

\item{outdir}{A character specifying an existing directory where output files
will be written to. Defaults to \code{"."}.}

\item{timeframe}{A character of either \code{"seasonal"} or \code{"complete"} specifying
if the calculation should be applied for the complete time series or for
each year independently. Defaults to \code{"complete"}. If \code{"seasonal"} is specified
the \code{years} variables cannot be left empty.}

\item{years}{A character vector specifying the individual years for the calculation
in case timeframe is \code{"seasonal"} in the form of e.g. \code{c("2016", "2017")}.}

\item{threads}{Number of cores used for parallel processing}
}
\value{
Nothing. However a \code{GTiff} and a \code{GPKG} with the zonal statistics are written to
  \code{outdir} for every resulting timestep.
}
\description{
This function is called by users to calculate a spectral index and extract zonal
statistics for a sf object containing \code{POLYGONS} or \code{MULTIPOLYGONS}. Any \code{ZM} dimensions
of the \code{sf} object are dropped before calculation. Users can specify the timeframe
of the calculation as \code{"seasonal"}. In this way, the calculation for each year
specified in the \code{years} variables is done separately. This allows users to only
cover specific timeframes within several years. If timeframe is \code{"complete"}
it is assumed that the complete timeseries of the input files should be covered
by the temporal resolution specified in the \code{dt} variable.
}
\note{
See \code{\link[gdalcubes:cube_view]{cube_view()}} and \code{\link[gdalcubes:zonal_statistics]{zonal_statistics()}} for more information.
}
\author{
Darius Görgen, Dr. Fabian Löw (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
