% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen2_processing.R
\name{calcZS}
\alias{calcZS}
\title{Calculate Zonal statistics with gdlcubes}
\usage{
calcZS(
  aoi,
  idcol,
  rasterfiles,
  dates,
  band_name,
  zonalstat,
  epsg,
  dx,
  dy,
  dt,
  aggregation,
  resampling,
  threads
)
}
\arguments{
\item{aoi}{A sf-object with POLYGONS or MULTIPOLYGONS. For objects with ZM
dimensions these dimensions are dropped before calculation.}

\item{idcol}{A column name of the \code{aoi} object uniquely identifying the polygons.
The values found in this columns will be appended to the resulting sf object
to allow feature identification.}

\item{rasterfiles}{Input raster files}

\item{dates}{Input dates}

\item{band_name}{Input band names}

\item{zonalstat}{Zonal stats to calculate}

\item{epsg}{A character specifying the epsg code for the the cube view specified
as \code{"EPSG:<code>"}. The resulting \code{GTiff} files and the \code{GPKG}
with the zonal statistics will be projected into this CRS.
\code{dx} and \code{dy} are understood in the units of the CRS specified here.
If left empty, the CRS of the \code{aoi} variable will be used.}

\item{dx}{A numeric specifying the resolution of the output raster in x dimension
expressed in the map unit specified in \code{epsg.} If \code{epsg} is left
empty it defaults to the map unit of \code{aoi.}}

\item{dy}{A numeric specifying the resolution of the output raster in y dimension
expressed in the map unit specified in \code{epsg}.
If \code{epsg} is left empty it defaults to the map unit of \code{aoi}.}

\item{dt}{Resolution of the time dimension expressed as ISO8601 string
(e.g. \code{"P16D"} for 16 days).}

\item{aggregation}{A character vector specifying the aggregation method for
images of the same time stamp. Must be one of \code{"min"}, \code{"max"},
\code{"mean"}, \code{"median"}, or \code{"first"}. Defaults to \code{"mean"}.}

\item{resampling}{A character specifying the resampling method used to transform
images from their native CRS to the output CRS. Can be any of the methods
supported by gdalwarp (see \url{www.gdal.org/programs/gdalwarp.html}).
Defaults to \code{"bilinear"}.}

\item{threads}{Number of cores used for parallel processing}
}
\value{
A \code{sf} object in long format with columns for the specified zonal statistics,
 and rows for each feature and input time.
}
\description{
Calculate Zonal statistics with gdlcubes
}
\author{
Darius Görgen, Dr. Fabian Löw (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
