% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen2_processing.R
\name{check_indices}
\alias{check_indices}
\title{Get available indices}
\usage{
check_indices()
}
\value{
A dataframe
}
\description{
This functions returns a dataframe with the available indices for calcualtion.
Use the values found in the column \code{name} to specify an individual index
}
\examples{
library(mapme.vegetation)
check_indices()[1:10,]
}
\author{
Darius Görgen, Dr. Fabian Löw (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} KfW FZ Evaluierung \email{FZ-Evaluierung@kfw.de}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} KfW Bankengruppe
\cr
\emph{License:} GPL-3
}
