% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_funs.R
\name{compare_epochs}
\alias{compare_epochs}
\title{Function to compare two epochs}
\usage{
compare_epochs(
  epoch1_files = NULL,
  epoch2_files = NULL,
  order = c(1, 2),
  aggregation = "mean",
  stat = "-"
)
}
\arguments{
\item{epoch1_files}{A character vector pointing to one or more raster files
to be considered as part of the first epoch.}

\item{epoch2_files}{character vector pointing to one or more raster files
to be considered as part of the second epoch.}

\item{order}{A numeric vector of length 2 indicating the order of the epoch
comparison. Either \code{c(1,2)} which will put the first epoch first or
\code{c(2,1)} which will put the second epoch first.}

\item{aggregation}{A character vector indicating the aggregation operations
applied to each pixel individually for each epoch when more than one file
per epoch are handed to the function. Must be one of mean, max, min, median,
sum, range, and prod.}

\item{stat}{A charachter vector indicating the arithmetic operation for the
comparison. Must be one of "-", "+", "*", "/".}
}
\value{
A spatRaster.
}
\description{
This function is used to compare the raster files of two different epochs.
The order as well as the arithmetic operation of the comparision can be
specified by the user. The paths to the raster files comprising the
first and second epoch are the main inputs. When more than one raster file
per epoch is handed to the function an aggregation function is applied per
pixel for each of the epochs.
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
