% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen2_processing.R
\name{createCloudMask}
\alias{createCloudMask}
\title{Cloud Mask raster for Sentinel 2 L1C}
\usage{
createCloudMask(safe, overwrite = T)
}
\arguments{
\item{safe}{A length one character vector pointing to the root
of a \code{SAFE} directory.}

\item{overwrite}{A logical indicating if an already existing cloud
mask raster should be overwritten. Defaults to \code{TRUE}}
}
\value{
Nothing. However, a \code{GTiff} file is written into the imagery
  directory of the \code{SAFE} product found at \code{GRANULE/<tile-id>/IMG_DATA/<tile-id>_CM.tif}.
}
\description{
This function is used to create a \code{GTiff} cloud mask from
the \code{gml} vectorized cloud mask product of Sentinel Level 1C data.
A \code{GTiff} file called \code{<tile-id>_CM.tif}
is written to \code{GRANULE/<tile-id>/IMG_DATA} in 60m resolution. A value
of 1 indicates the presence of a cloud while a value of 0 represents
clear sky conditions according to the cloud product.
}
\author{
Darius Görgen, Dr. Fabian Löw (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
