% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sen2_processing.R
\name{cube_indexcalc}
\alias{cube_indexcalc}
\title{Calculate a spectral index with gdalcubes}
\usage{
cube_indexcalc(
  safes,
  aoi,
  epsg = NULL,
  dx = NULL,
  dy = NULL,
  dt = NULL,
  aggregation = "mean",
  resampling = "bilinear",
  tmpagg = FALSE,
  index = NULL,
  label = "",
  outdir = ".",
  threads
)
}
\arguments{
\item{safes}{A character vector with the \code{.jp2} and \code{.tif} files of the
selected \code{SAFE} directories. Obtained with \code{list.files(<safedir>, "jp2$|.tif$", recursive =T, full.names = T)}}

\item{aoi}{A sf-object with POLYGONS or MULTIPOLYGONS. For objects with ZM
dimensions these dimensions are dropped before calculation.}

\item{epsg}{A character specifying the epsg code for the the cube view specified
as \code{"EPSG:<code>"}. The resulting \code{GTiff} files and the \code{GPKG}
with the zonal statistics will be projected into this CRS.
\code{dx} and \code{dy} are understood in the units of the CRS specified here.
If left empty, the CRS of the \code{aoi} variable will be used.}

\item{dx}{A numeric specifying the resolution of the output raster in x dimension
expressed in the map unit specified in \code{epsg.} If \code{epsg} is left
empty it defaults to the map unit of \code{aoi.}}

\item{dy}{A numeric specifying the resolution of the output raster in y dimension
expressed in the map unit specified in \code{epsg}.
If \code{epsg} is left empty it defaults to the map unit of \code{aoi}.}

\item{dt}{Resolution of the time dimension expressed as ISO8601 string
(e.g. \code{"P16D"} for 16 days).}

\item{aggregation}{A character vector specifying the aggregation method for
images of the same time stamp. Must be one of \code{"min"}, \code{"max"},
\code{"mean"}, \code{"median"}, or \code{"first"}. Defaults to \code{"mean"}.}

\item{resampling}{A character specifying the resampling method used to transform
images from their native CRS to the output CRS. Can be any of the methods
supported by gdalwarp (see \url{www.gdal.org/programs/gdalwarp.html}).
Defaults to \code{"bilinear"}.}

\item{tmpagg}{A logical indicating if temporal aggregation of all timestemps should
be performed calling \code{\link[gdalcubes:reduce_time]{reduce_time()}}. Defaults to \code{FALSE}.}

\item{index}{A character specifying the short name of the index to be calculated.
Available indices can be checked with \code{check_indices()}.}

\item{label}{A character label which is appended to the output files in the
form of \code{<label>_<index>_<timestep>}. Files with the same name in
\code{outdir} will be overwritten without warning.}

\item{outdir}{A character specifying an existing directory where output files
will be written to. Defaults to \code{"."}.}

\item{threads}{Number of cores used for parallel processing}

\item{zonalstat}{A character specifying the aggregation statistic to be used by
\code{\link[gdalcubes:zonal_statistics]{zonal_statistics()}}.
Multiple selections are supported. Must be included in \code{"min"}, \code{"max"},
\code{"mean"}, \code{"median"},
\code{"count"}, \code{"sum"}, \code{"prod"}, \code{"var"}, and \code{"sd"}.
Defaults to \code{"mean"}.}
}
\value{
Nothing. However a \code{GTiff} and a \code{GPKG} with the zonal statistics are written to
  \code{outdir} for every resulting timestep.
}
\description{
This function is used internally to calculate a spectral index using the
gdalcubes package. It is called by \code{calcIndex()} on a given set of \code{SAFE}
directories. A user defined cube view defined by the x and y resolution,
the time resolution, a CRS, as well as aggregation methods is applied.
Additionally, the time dimension can be aggregated which is useful for seasonal
calculations of a given index. The output of the function is a \code{GTiff} file
for each time step labeled as \code{<label>_<index>_<timestep>.tif}. Zonal statistics
are written to a Geopackage for each timestep labeled as \code{<label>_<index>_<timestep>.gpkg}
to outdir.
}
\note{
See \code{\link[gdalcubes:cube_view]{cube_view()}} and \code{\link[gdalcubes:zonal_statistics]{zonal_statistics()}} for more information.
}
\author{
Darius Görgen, Dr. Fabian Löw (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
\keyword{internal}
