% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_cube.r
\name{.calc_index}
\alias{.calc_index}
\title{Calculate a spectral index with gdalcubes}
\usage{
.calc_index(
  files,
  format = "mapme.vegetation/inst/sentinel2_l2a_cog.json",
  DB_path = NULL,
  dx = NULL,
  dy = NULL,
  dt = NULL,
  srs = NULL,
  bbox = NULL,
  aggregation = "median",
  resampling = "bilinear",
  after = NULL,
  before = NULL,
  tmpagg = FALSE,
  index = NULL,
  bands = NULL,
  label = NULL,
  outdir = ".",
  mask_layer = "SCL",
  mask_values = c(3, 8, 9),
  mask_invert = FALSE,
  overwrite = F,
  chunking = c(1, 256, 256),
  threads = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{files}{A character vector with the filepaths to the raster files for which
gap-filling and a smoothing function shall be applied.}

\item{format}{A charachter vector pointing to a json file with a gdalcubes format
description of the input files.}

\item{DB_path}{A charachter vector where the gdalcubes data base file indexing
the spatiotemporal extents of files. Defaults to NULL which means a temporary
file is used. Can be pointed to a file ending in ".db" to speed up later iterations.}

\item{dx}{A numeric specifying the resolution of the output raster in x dimension
expressed in the map unit specified in \code{epsg.} If \code{epsg} is left
empty it defaults to the map unit of \code{aoi.}}

\item{dy}{A numeric specifying the resolution of the output raster in y dimension
expressed in the map unit specified in \code{epsg}.
If \code{epsg} is left empty it defaults to the map unit of \code{aoi}.}

\item{dt}{Resolution of the time dimension expressed as ISO8601 string
(e.g. \code{"P16D"} for 16 days).}

\item{srs}{Target spatial reference system as a string; can be a proj4 definition,
WKT, or in the form "EPSG:XXXX" (see \code{\link[gdalcubes]{cube_view}})}

\item{bbox}{A numeric vector of lenght four indicating the spatial bounding
box of the query (xmin, ymin, xmax, ymax) in geographic coordinates. Defaults
to `NULL` which will set the spatial extent to global, i.e. \code{c(-180, -90, 180, 90)}.}

\item{aggregation}{A character vector specifying the aggregation method for
images of the same time stamp. Must be one of \code{"min"}, \code{"max"},
\code{"mean"}, \code{"median"}, or \code{"first"}. Defaults to \code{"median"}.}

\item{resampling}{A character specifying the resampling method used to transform
images from their native CRS to the output CRS. Can be any of the methods
supported by gdalwarp (see \url{www.gdal.org/programs/gdalwarp.html}).
Defaults to \code{"bilinear"}.}

\item{after}{A length one character vector in the form `%Y-%m-%d` specifying
the earliest date (inclusive) included in the temporal extent.}

\item{before}{A length one character vector in the form `%Y-%m-%d` specifying
the latest date (inclusive) included in the temporal extent.}

\item{tmpagg}{A logical indicating if temporal aggregation of all timestemps should
be performed calling \code{\link[gdalcubes]{reduce_time}}. Defaults to \code{FALSE}.}

\item{index}{A character specifying the short name of the index to be calculated.
Available indices can be checked with \code{check_indices()}. Defaults
to NULL. When indices are calculated argument bands must be set to NULL.}

\item{bands}{A character vector specifying the bands to extract. Defaults to
NULL. If specified argument index must be set to NULL.}

\item{label}{A character label which is appended to the output files in the
form of \code{<label>_<index>_<timestep>}. Files with the same name in
\code{outdir} will be overwritten without warning.}

\item{outdir}{A character specifying an existing directory where output files
will be written to. Defaults to \code{"."}.}

\item{mask_layer}{Name of a layer in the data cube used for masking.}

\item{mask_values}{An integer vector with mask values. Masking behaviour
is controlled with the \code{mask_invert} argument.}

\item{mask_invert}{If TRUE the values specified in \code{mask_values} are
considered valid and pixels with these values are kept. If FALSE pixel
with these values are masked out. Default is FALSE.}

\item{overwrite}{A logical indicating if existing files in outdir should be
overwritten.}

\item{chunking}{Vector of length 3 defining the size of data cube chunks in the
order time, y, x. (see \code{\link[gdalcubes]{raster_cube}})}

\item{threads}{Number of cores used for parallel processing}

\item{verbose}{A logical indicating the verbosity.}

\item{...}{additional arguments to  \code{\link[gdalcubes]{write_tif}}}
}
\value{
Nothing. However a \code{GTiff} and a \code{GPKG} with the zonal statistics are written to
  \code{outdir} for every resulting timestep.
}
\description{
This function is used internally to calculate a spectral index using the
gdalcubes package. It is called by \code{calcIndex()} on a given set of \code{SAFE}
directories. A user defined cube view defined by the x and y resolution,
the time resolution, a CRS, as well as aggregation methods is applied.
Additionally, the time dimension can be aggregated which is useful for seasonal
calculations of a given index. The output of the function is a \code{GTiff} file
for each time step labeled as \code{<label>_<index>_<timestep>.tif}. Zonal statistics
are written to a Geopackage for each timestep labeled as \code{<label>_<index>_<timestep>.gpkg}
to outdir.
}
\note{
See \code{\link[gdalcubes]{cube_view}} and \code{\link[gdalcubes]{zonal_statistics}} for more information.
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
\keyword{internal}
