% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.check_collection}
\alias{.check_collection}
\title{Check availability of a collection}
\usage{
.check_collection(s, collection)
}
\arguments{
\item{s}{A STAC endpoint created with `rstac::stac()`.}

\item{collection}{A length one charachter vector indicating a collection name}
}
\value{
Nothing. The function is used to check the availablility of a collection.
}
\description{
This function checks if a user specified collection is available at the STAC
endpoint and stops with an error message if that is not the case
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
\keyword{internal}
