% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.check_dates}
\alias{.check_dates}
\title{Check specified dates and arrange for download}
\usage{
.check_dates(after, before, timeframe)
}
\arguments{
\item{after}{A length one character vector in the form `%Y-%m-%d` specifying
the earliest date (inclusive) included in the temporal extent.}

\item{before}{A length one character vector in the form `%Y-%m-%d` specifying
the latest date (inclusive) included in the temporal extent.}

\item{timeframe}{A length one character vector indicating whether the query
should return the complete time series between `after` and `before` or
if the query should be based on seasons. Possbile values are `full` and `seasonal`.
Defaults to `full`.}
}
\value{
A data-frame with one row for `timeframe = "full"` and multiple rows
  depending on the number of years between `after` and `before` if `timeframe = "seasonal"`.
}
\description{
This function is used to check if the user specified dates can be coerced
to dates. Additionally, it arranges the dates based on the user wants a full
time series or seasonal periods between different years.
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
\keyword{internal}
