% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.check_smoother}
\alias{.check_smoother}
\title{Internal function used to sample the smoothing function}
\usage{
.check_smoother(cube, smoothing_function, n = 100)
}
\arguments{
\item{cube}{A data cube from which a random sample is drawn.}

\item{smoothing_function}{A smoothing function to be applied.}

\item{p}{A ratio of random pixels to be sampled.}
}
\value{
Nothing. Throws an error if the smoothing function does not
  maintain cube dimensions and removes NA's.
}
\description{
This function samples a ratio of p pixels of the data cube and applies
the smoothing function on them. The output is checked to contain no NA's
and an equal number of rows (bands) and columns (time steps) as the input
data cube.
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
\keyword{internal}
