% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.check_times}
\alias{.check_times}
\title{Internal function to check date vectors}
\usage{
.check_times(times)
}
\arguments{
\item{times}{A charachter vector representing dates}
}
\value{
Nothing
}
\description{
Function used across the package to check the input of date vectors.
Throws errors when times is not a character vector and if not all
elements can be coerced to dates.
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
\keyword{internal}
