% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.download_assets}
\alias{.download_assets}
\title{Download items from a STAC catalog}
\usage{
.download_assets(
  items = NULL,
  assets = NULL,
  outdir = ".",
  use_aria = FALSE,
  aria_bin = NULL,
  skip_existing = FALSE,
  max_concurrent = 4,
  max_connections = 4,
  retry = 2,
  verbose = TRUE
)
}
\arguments{
\item{items}{A list of items returned by `rstac::stac_search()`}

\item{assets}{Default is NULL. A possible charachter vector to specify which
assets to download. If assets do not match the function returns an error
and lists the available assets.}

\item{outdir}{A character vector of an existing directory. Defaults to the
current working directory.}

\item{use_aria}{A logical if the aria2 downloader shall be used. Defaultes to
FALSE which will download the items sequentially via GET requests. If TRUE
it is expected that the argument `aria_bin` points towards a valid
aria binary installation.}

\item{aria_bin}{A charachter vector pointing to a valid aria binary.}

\item{skip_existing}{A logical indicating if existing files should be removed from
the download list prior to downloading with aria. Might speed up the process
when a prior download was interrupted, however, when files were not downloaded
completley their download is not ressumed (check for .aria files in outdir).
Defaults to FALSE which means that existing files are not skipped.}

\item{max_concurrent}{Maximum number of connections to the server to load
a single file. Defaults to 4.}

\item{max_connections}{Maximum number of parallel connections to the server
to load files simoltaniously. Defaults to 4.}

\item{retry}{Integer. Maximum integer of retries for erroneous downloads.}

\item{verbose}{A logical indicating the verbosity.}
}
\value{
Nothing. Downloads specified assets into `outdir`.
}
\description{
This function is used internally to download a given list of items matching
a spatiotemporal query.
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
\keyword{internal}
