% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.query_items}
\alias{.query_items}
\title{Queries items from a STAC endpoint}
\usage{
.query_items(s, collection, bbox, datetime, max.items, verbose, max.cloud)
}
\arguments{
\item{s}{A STACconnection object from the \code{rstac} package.}

\item{collection}{A length one character vector indicating the name of the
collection to query for items.}

\item{bbox}{A numeric vector of length four indicating the spatial bounding
box of the query \code{(xmin, ymin, xmax, ymax)} in geographic coordinates. Defaults
to `NULL` which will set the spatial extent to global, i.e. \code{c(-180, -90, 180, 90)}.}

\item{datetime}{A length one character vector indicating the start and endpoint
of the temporal query.}

\item{max.items}{A length one integer indicating the maximum number of items
to return from a single query. Defaults to 500. Can be used for cases where
large queries will lead to long query times. But if that is the case it is
advisable to break up into small sub-queries so that the complete data can
be downloaded.}

\item{verbose}{A  logical indicating the verbosity.}

\item{max.cloud}{A length one integer indicating the maximum number of cloud
cover in percent to qualify a scene for download. Defaults to \code{100} which will
download all scenes matching the spatiotemporal extent.}
}
\value{
A items list object of \code{rstac}.
}
\description{
This function queries all items of a given collection of a STAC endpoint
matching a spatiotemporal extent.
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
\keyword{internal}
