% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.update_ext}
\alias{.update_ext}
\title{Internal function to update a spatio temporal extent}
\usage{
.update_ext(col, bbox = NULL, after = NULL, before = NULL)
}
\arguments{
\item{col}{A image collection}

\item{bbox}{A spatial bounding box}

\item{after}{A date character}

\item{before}{A date character}
}
\description{
Internal function to update a spatio temporal extent
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
\keyword{internal}
