% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_aws.R
\name{download_aws}
\alias{download_aws}
\title{Download S2 L2A COGs from AWS servers}
\usage{
download_aws(
  collection = "sentinel-s2-l2a-cogs",
  assets = NULL,
  bbox = NULL,
  after = NULL,
  before = NULL,
  timeframe = "full",
  max.cloud = 100,
  max.items = 500,
  outdir = ".",
  overwrite = FALSE,
  use_aria = F,
  aria_bin = NULL,
  skip_existing = FALSE,
  max_concurrent = 4,
  max_connections = 4,
  retry = 2,
  verbose = T,
  query_only = F
)
}
\arguments{
\item{collection}{A name of a collection. Defaults to \code{"sentinel-s2-l2a-cogs"},
currently the only supported collection.}

\item{assets}{A character vector indicating the assets to download, e.g. bands.
Defaults to \code{NULL} which will download all available assets.}

\item{bbox}{A numeric vector of lenght four indicating the spatial bounding
box of the query \code{c(xmin, ymin, xmax, ymax)} in geographic coordinates. Defaults
to \code{NULL} which will set the spatial extent to global, i.e. \code{c(-180, -90, 180, 90)}.}

\item{after}{A length one character vector in the form `%Y-%m-%d` specifying
the earliest date (inclusive) included in the temporal extent.}

\item{before}{A length one character vector in the form `%Y-%m-%d` specifying
the latest date (inclusive) included in the temporal extent.}

\item{timeframe}{A length one character vector indicating whether the query
should return the complete time series between \code{after} and \code{before} or
if the query should be based on seasons. Possbile values are \code{"full"} and \code{"seasonal"}.
Defaults to \code{"full"}.}

\item{max.cloud}{A length one integer indicating the maximum number of cloud
cover in percent to qualify a scene for download. Defaults to \code{"100"} which will
download all scenes matching the spatiotemporal extent.}

\item{max.items}{A length one integer indicating the maximum number of items
to return from a single query. Defaults to 500. Can be used for cases where
large queries will lead to long query times. But if that is the case it is
advisable to break up into small sub-queries so that the complete data can
be downloaded.}

\item{outdir}{A character vector of an existing directory. Defaults to the
current working directory. Throws an error if the directory does not exist.}

\item{overwrite}{A logical indicator indicating if present data sets in \code{outdir}
should be overwritten. Defaults to \code{FALSE}.}

\item{use_aria}{A logical if the aria2 downloader shall be used. Defaultes to
FALSE which will download the items sequentially via GET requests. If TRUE
it is expected that the argument \code{aria_bin} points towards a valid
aria binary installation.}

\item{aria_bin}{A charachter vector pointing to a valid aria binary.}

\item{skip_existing}{A logical indicating if existing files should be removed from
the download list prior to downloading with aria. Might speed up the process
when a prior download was interrupted, however, when files were not downloaded
completley their download is not ressumed (check for .aria files in outdir).
Defaults to FALSE which means that existing files are not skipped.}

\item{max_concurrent}{Maximum number of connections to the server to load
a single file. Defaults to \code{4}.}

\item{max_connections}{Maximum number of parallel connections to the server
to load files simoltaniously. Defaults to 4.}

\item{retry}{Integer. Maximum integer of retries for erroneous downloads.}

\item{verbose}{A logical indicating the verbosity.}

\item{query_only}{Logical. Defaults to FALSE which means matching items are downloaded.
Else matching items are returned in a list object.}
}
\value{
Nothing. This function downloads matching assets of the spatiotemporal
  query to the specified directory in \code{outdir}.
}
\description{
This function can be used for a spatiotemporal query of S2 L2A data on AWS
servers and downloads either all or specific assets found at the STAC API
endpoint. Internally the function depends on the `rstac` package which
makes queries to STAC endpoints very convinient. Note, currently only S2 L2A
COGs can be downloaded using this function.
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
