% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_funs.R
\name{extract_pixels}
\alias{extract_pixels}
\title{Extract pixel values of polygons from raster files}
\usage{
extract_pixels(files = NULL, bands = NULL, times = NULL, aoi = NULL)
}
\arguments{
\item{files}{A chachter vector indicating the raster files.}

\item{bands}{A charachter vector indicating the names of bands for each raster files.}

\item{times}{A charachter vector indicating the date for each raster file}

\item{aoi}{An sf object of polygons for which to extract the pixel values}
}
\value{
A dataframe object with the extracted value amended by the metadata
  of the aoi object.
}
\description{
This function extracts all pixel values within polygons based on a time series
of rasters. Users are expected to specify the name vector of bands as well
as a vector indicating the date for each raster.
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
