% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_funs.R
\name{extract_zonalstats}
\alias{extract_zonalstats}
\title{Calculate Zonal statistics with gdalcubes}
\usage{
extract_zonalstats(
  aoi = NULL,
  idcol = NULL,
  files = NULL,
  times,
  bands,
  zonalfuns,
  bbox,
  after,
  before,
  srs,
  dx,
  dy,
  dt,
  aggregation,
  resampling,
  outpath = NULL,
  threads = 1,
  ...
)
}
\arguments{
\item{aoi}{A sf object with polygons / multipolygons}

\item{idcol}{A column name of the \code{aoi} object uniquely identifying the polygons.
The values found in this columns will be appended to the resulting sf object
to allow feature identification.}

\item{files}{Input raster files}

\item{times}{Input dates}

\item{bands}{Input band names}

\item{zonalfuns}{A named list with functions for zonal statistics as elements}

\item{bbox}{Bounding Box}

\item{after}{Date after}

\item{before}{Date before}

\item{srs}{target spatial reference system as a string; can be a proj4 definition, WKT, or in the form "EPSG:XXXX"}

\item{dx}{resolution x}

\item{dy}{resolution y}

\item{dt}{resoltion time}

\item{aggregation}{aggregation method as string, defining how to deal with pixels containing data from multiple images, can be "min", "max", "mean", "median", or "first"}

\item{resampling}{method used in gdalwarp when images are read, can be "near", "bilinear", "bicubic" or others as supported by gdalwarp (see https://gdal.org/programs/gdalwarp.html)}

\item{outpath}{Path for the output GeoPakackage}

\item{threads}{Number of threads}

\item{...}{Additional arguments to st_write}
}
\value{
A \code{sf} object in long format with columns for the specified zonal statistics,
 and rows for each feature and input time.
}
\description{
Calculate Zonal statistics with gdalcubes
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
