% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_cube.r
\name{index_cube}
\alias{index_cube}
\title{Spectral index and zonal statistics}
\usage{
index_cube(
  files = NULL,
  format = system.file("sentinel2_l2a_cog.json", package = "mapme.vegetation"),
  DB_path = NULL,
  dx = NULL,
  dy = NULL,
  dt = NULL,
  srs = NULL,
  bbox = NULL,
  after = NULL,
  before = NULL,
  timeframe = "full",
  aggregation = "median",
  resampling = "bilinear",
  tmpagg = FALSE,
  index = NULL,
  bands = NULL,
  label = "",
  outdir = ".",
  overwrite = F,
  mask_layer = NULL,
  mask_values = NULL,
  mask_invert = FALSE,
  chunking = c(1, 256, 256),
  threads = 1,
  ...
)
}
\arguments{
\item{files}{A character vector with the filepaths to the raster files for which
gap-filling and a smoothing function shall be applied.}

\item{format}{A charachter vector pointing to a json file with a gdalcubes format
description of the input files.}

\item{DB_path}{A charachter vector where the gdalcubes data base file indexing
the spatiotemporal extents of files. Defaults to NULL which means a temporary
file is used. Can be pointed to a file ending in ".db" to speed up later iterations.}

\item{dx}{A numeric specifying the resolution of the output raster in x dimension
expressed in the map unit specified in \code{epsg.} If \code{epsg} is left
empty it defaults to the map unit of \code{aoi.}}

\item{dy}{A numeric specifying the resolution of the output raster in y dimension
expressed in the map unit specified in \code{epsg}.
If \code{epsg} is left empty it defaults to the map unit of \code{aoi}.}

\item{dt}{Resolution of the time dimension expressed as ISO8601 string
(e.g. \code{"P16D"} for 16 days).}

\item{srs}{Target spatial reference system as a string; can be a proj4 definition,
WKT, or in the form "EPSG:XXXX" (see \code{\link[gdalcubes]{cube_view}})}

\item{bbox}{A numeric vector of lenght four indicating the spatial bounding
box of the query (xmin, ymin, xmax, ymax) in geographic coordinates. Defaults
to `NULL` which will set the spatial extent to global, i.e. \code{c(-180, -90, 180, 90)}.}

\item{after}{A length one character vector in the form `%Y-%m-%d` specifying
the earliest date (inclusive) included in the temporal extent.}

\item{before}{A length one character vector in the form `%Y-%m-%d` specifying
the latest date (inclusive) included in the temporal extent.}

\item{timeframe}{A character of either \code{"seasonal"} or \code{"complete"} specifying
if the calculation should be applied for the complete time series or for
each year independently. Defaults to \code{"complete"}. If \code{"seasonal"} is specified
the \code{years} variables cannot be left empty.}

\item{aggregation}{A character vector specifying the aggregation method for
images of the same time stamp. Must be one of \code{"min"}, \code{"max"},
\code{"mean"}, \code{"median"}, or \code{"first"}. Defaults to \code{"median"}.}

\item{resampling}{A character specifying the resampling method used to transform
images from their native CRS to the output CRS. Can be any of the methods
supported by gdalwarp (see \url{www.gdal.org/programs/gdalwarp.html}).
Defaults to \code{"bilinear"}.}

\item{tmpagg}{A logical indicating if temporal aggregation of all timestemps should
be performed calling \code{\link[gdalcubes]{reduce_time}}. Defaults to \code{FALSE}.}

\item{index}{A character specifying the short name of the index to be calculated.
Available indices can be checked with \code{check_indices()}. Defaults
to NULL. When indices are calculated argument bands must be set to NULL.}

\item{bands}{A character vector specifying the bands to extract. Defaults to
NULL. If specified argument index must be set to NULL.}

\item{label}{A character label which is appended to the output files in the
form of \code{<label>_<index>_<timestep>}. Files with the same name in
\code{outdir} will be overwritten without warning.}

\item{outdir}{A character specifying an existing directory where output files
will be written to. Defaults to \code{"."}.}

\item{overwrite}{A logical indicating if existing files in outdir should be
overwritten.}

\item{mask_layer}{Name of a layer in the data cube used for masking.}

\item{mask_values}{An integer vector with mask values. Masking behaviour
is controlled with the \code{mask_invert} argument.}

\item{mask_invert}{If TRUE the values specified in \code{mask_values} are
considered valid and pixels with these values are kept. If FALSE pixel
with these values are masked out. Default is FALSE.}

\item{chunking}{Vector of length 3 defining the size of data cube chunks in the
order time, y, x. (see \code{\link[gdalcubes]{raster_cube}})}

\item{threads}{Number of cores used for parallel processing}

\item{...}{additional arguments to  \code{\link[gdalcubes]{write_tif}}}
}
\value{
Nothing. However a \code{GTiff} and a \code{GPKG} with the zonal statistics are written to
  \code{outdir} for every resulting timestep.
}
\description{
This function is called by users to calculate a spectral index and extract zonal
statistics for a sf object containing \code{POLYGONS} or \code{MULTIPOLYGONS}. Any \code{ZM} dimensions
of the \code{sf} object are dropped before calculation. Users can specify the timeframe
of the calculation as \code{"seasonal"}. In this way, the calculation for each year
specified in the \code{years} variables is done separately. This allows users to only
cover specific timeframes within several years. If timeframe is \code{"complete"}
it is assumed that the complete timeseries of the input files should be covered
by the temporal resolution specified in the \code{dt} variable.
}
\note{
See \code{\link[gdalcubes]{cube_view}} and \code{\link[gdalcubes]{zonal_statistics}} for more information.
}
\author{
Darius Görgen, Dr. Fabian Löw (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
