% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_scl.R
\name{scl_buffer}
\alias{scl_buffer}
\title{Apply a buffer to SCL cloud mask for specific values}
\usage{
scl_buffer(scl_files, mask_values, mask_buffer, grass_bin, outdir, threads)
}
\arguments{
\item{scl_files}{A chrachter vector to the input SCL files.}

\item{mask_values}{An integer vector indiacating the values of the raster
which are considered as values to mask. Note that the resulting output
raster will get the first value for all pixels within the spatial buffer.}

\item{mask_buffer}{A numeric indicating the distance of the spatial buffer
in map units (usuall meters).}

\item{grass_bin}{A charachter vector to the directory of the GRASS binaries.}

\item{outdir}{A charachter vector pointin towards a directory where the output
files should be written to. Note that this should be a different directory
from where the original files are stored because the output files will have
identical names.}

\item{threads}{The number of threads for parallel computation.}
}
\value{
Nothing. Writes buffered SCL mask to outdir. All cells within the
  buffer will obtain the value of the first element in mask_values.
}
\description{
This function allows to apply spatial buffers for SCL tiles in order to enrich
the cloud cover classification and exclude potentially degradated pixels.
It uses GRASS GIS for fast processing. User's will need to provide a working
GRASS installation in order to use this function. It can be applied in parallel
to further increase computation speed.
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
