% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_cube.R
\name{smooth_cube}
\alias{smooth_cube}
\title{Function to fill gaps and smooth a time series of vegetation indices}
\usage{
smooth_cube(
  files = NULL,
  bands = NULL,
  times = NULL,
  dx = NULL,
  dy = NULL,
  dt = NULL,
  srs = NULL,
  after = NULL,
  before = NULL,
  timeframe = "full",
  bbox = NULL,
  aggregation = "mean",
  resampling = "bilinear",
  smoothing_function = NULL,
  chunking = c(1, 256, 256),
  threads = 1,
  outdir = ".",
  overwrite = F,
  label = "",
  n = 100,
  verbose = T,
  ...
)
}
\arguments{
\item{files}{A character vector with the filepaths to the raster files for which
gap-filling and a smoothing function shall be applied.}

\item{bands}{A character vector indicating the names of the bands in each
of the raster files specified in the \code{files} argument.
(see \code{\link[gdalcubes]{create_image_collection}})}

\item{times}{A character vector indicating the date of each raster file specified
in the \code{files} argument in the form of "%Y%m%d".
(see \code{\link[gdalcubes]{create_image_collection}})}

\item{dx}{The spatial resolution of the outcome raster files in the \code{x} dimension
in the units of the coordinate reference system specified witht the \code{srs} argument.
(see \code{\link[gdalcubes]{cube_view}})}

\item{dy}{The spatial resolution of the outcome raster files in the \code{y} dimension
in the units of the coordinate reference system specified witht the \code{srs} argument.
(see \code{\link[gdalcubes]{cube_view}})}

\item{dt}{The temporal resolution of the outcome raster files.
(see \code{\link[gdalcubes]{cube_view}})}

\item{srs}{Target spatial reference system as a string; can be a proj4 definition,
WKT, or in the form "EPSG:XXXX" (see \code{\link[gdalcubes]{cube_view}})}

\item{after}{A length one character vector in the form "%Y-%m-%d" specifying
the earliest date (inclusive) included in the temporal extent.}

\item{before}{A length one character vector in the form "%Y-%m-%d" specifying
the latest date (inclusive) included in the temporal extent.}

\item{timeframe}{A character of either \code{"seasonal"} or \code{"complete"} specifying
if the calculation should be applied for the complete time series or for
each year independently. Defaults to \code{"complete"}. If \code{"seasonal"} is specified
the \code{years} variables cannot be left empty.}

\item{bbox}{A numeric vector of lenght four indicating the spatial bounding
box of the query (xmin, ymin, xmax, ymax) in geographic coordinates. Defaults
to `NULL` which will set the spatial extent to global, i.e. \code{c(-180, -90, 180, 90)}.}

\item{aggregation}{aggregation method as string, defining how to deal with pixels
containing data from multiple images, can be "min", "max", "mean", "median", or "first"
(see \code{\link[gdalcubes]{cube_view}})}

\item{resampling}{resampling method used in gdalwarp when images are read, can be
"near", "bilinear", "bicubic" or others as supported by gdalwarp
(see https://gdal.org/programs/gdalwarp.html and \code{\link[gdalcubes]{cube_view}})}

\item{smoothing_function}{Defaults to NULL which means that a linear
interpolation coupled with \code{\link[signal]{sgolayfilt}}` at its standard
parametrization is applied. Can be substituted by a function.
This function is applied across the time dimension of the data cube and imputes
NA's. The return vector must be of equal length to the time dimension, must
return one row for each band in the data cube and must not include any NA's.
See the example for the standard function which is applied when the \code{smoothing_function}
is omitted.}

\item{chunking}{Vector of length 3 defining the size of data cube chunks in the
order time, y, x. (see \code{\link[gdalcubes]{raster_cube}})}

\item{threads}{Number of threads used to process data cubes
(see \code{link[gdalcubes]{gdalcubes_options}})}

\item{outdir}{A character vector pointing to an existing directory where
the output rasters are written to.}

\item{overwrite}{A logical indicating if existing files in outdir should be
overwritten.}

\item{label}{A character vector which is used as a prefix to the output filenames.}

\item{n}{The ratio of pixels for which a custom \code{smoothing_function} is tested
on randomly selected pixels to test if it is applicable. Defaults to \code{0.01}.}

\item{verbose}{A logical indicating the verbosity.}

\item{...}{additional arguments to  \code{\link[gdalcubes]{write_tif}}}
}
\description{
This function is used to conduct a linear interpolation of missing values
in the time series of vegetation indices on a pixel basis. By default
a Savitzkiy-Golay filter is used to smooth the time series. A custom
function for the interpolation and smoothing can be specified. It can be used
on equally spaced data cubes in the time dimension where missing values, e.g.
through clouds, are an issue. The result of the function is a dense time series
where missing values are imputed and a smoothing function is applied to reduce
the noise in the signal.
}
\examples{
\dontrun{
smoothing_function = function(x){
results = lapply(1:nrow(x), function(i) {
 y = x[i,]
 lin = zoo::na.approx(y, na.rm = F, rule = 2)
 svg = signal::sgolayfilt(lin)
 svg
 })
do.call(rbind, results)
}
}
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
