% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_cube.R
\name{trend_cube}
\alias{trend_cube}
\title{Compute a trend raster on a raster time series}
\usage{
trend_cube(
  files = NULL,
  bands = NULL,
  times = NULL,
  dx = NULL,
  dy = NULL,
  dt = NULL,
  srs = NULL,
  after = NULL,
  before = NULL,
  bbox = NULL,
  aggregation = "median",
  resampling = "bilinear",
  trend_function = NULL,
  filter_p = NULL,
  chunking = c(1, 256, 256),
  threads = 1,
  outdir = ".",
  label = NULL,
  verbose = T
)
}
\arguments{
\item{files}{A character vector with the filepaths to the raster files for which
gap-filling and a smoothing function shall be applied.}

\item{bands}{A character vector indicating the names of the bands in each
of the raster files specified in the \code{files} argument.
(see \code{\link[gdalcubes]{create_image_collection}})}

\item{times}{A character vector indicating the date of each raster file specified
in the \code{files} argument in the form of "%Y%m%d".
(see \code{\link[gdalcubes]{create_image_collection}})}

\item{dx}{The spatial resolution of the outcome raster files in the \code{x} dimension
in the units of the coordinate reference system specified witht the \code{srs} argument.
(see \code{\link[gdalcubes]{cube_view}})}

\item{dy}{The spatial resolution of the outcome raster files in the \code{y} dimension
in the units of the coordinate reference system specified witht the \code{srs} argument.
(see \code{\link[gdalcubes]{cube_view}})}

\item{dt}{The temporal resolution of the outcome raster files.
(see \code{\link[gdalcubes]{cube_view}})}

\item{srs}{Target spatial reference system as a string; can be a proj4 definition,
WKT, or in the form "EPSG:XXXX" (see \code{\link[gdalcubes]{cube_view}})}

\item{after}{A length one character vector in the form "%Y-%m-%d" specifying
the earliest date (inclusive) included in the temporal extent.}

\item{before}{A length one character vector in the form "%Y-%m-%d" specifying
the latest date (inclusive) included in the temporal extent.}

\item{bbox}{A numeric vector of lenght four indicating the spatial bounding
box of the query (xmin, ymin, xmax, ymax) in geographic coordinates. Defaults
to `NULL` which will set the spatial extent to global, i.e. \code{c(-180, -90, 180, 90)}.}

\item{aggregation}{aggregation method as string, defining how to deal with pixels
containing data from multiple images, can be "min", "max", "mean", "median", or "first"
(see \code{\link[gdalcubes]{cube_view}})}

\item{resampling}{resampling method used in gdalwarp when images are read, can be
"near", "bilinear", "bicubic" or others as supported by gdalwarp
(see https://gdal.org/programs/gdalwarp.html and \code{\link[gdalcubes]{cube_view}})}

\item{trend_function}{Defaults to NULL which means that \code{\link[trend]{mk.test}}
with its standard parametrization is applied. Can be substituted by a function.
This function is applied across the time dimension of the data cube and imputes
NA's. The return vector must be of equal length to the time dimension, must
return one row for each band in the data cube and must not include any NA's.
See the example for the standard function which is applied when the \code{smoothing_function}
is omitted.}

\item{filter_p}{Defaults to FALSE, that means pixels are not filtered based on
the p-value. Can be set to a numeric value below 1 to mask keep all values
with a p-value smaller \code{p}.}

\item{chunking}{Vector of length 3 defining the size of data cube chunks in the
order time, y, x. (see \code{\link[gdalcubes]{raster_cube}})}

\item{threads}{Number of threads used to process data cubes
(see \code{link[gdalcubes]{gdalcubes_options}})}

\item{outdir}{A character vector pointing to an existing directory where
the output rasters are written to.}

\item{label}{A character vector which is used as a prefix to the output filenames.}

\item{verbose}{A logical indicating the verbosity.}
}
\description{
This function is used to calculate the estimate and a p-value of a trend
through a raster cube time series.
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
