% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{update_indexDB}
\alias{update_indexDB}
\title{Helper function to update tile index}
\usage{
update_indexDB(use_db = TRUE, force_update = F)
}
\arguments{
\item{use_db}{A logical parameter indicating if the tile index
should be saved locally as an SQLite database. This significantly
speeds up future queries but also increases the local disk
space used (from about 2 GB to 5 GB). Default is set to True.}

\item{force_update}{A logical parameter indicating that the
latest version of the tile index should be downloaded from
the Google Cloud.}
}
\value{
Nothing.
}
\description{
This function is a helper functions which downloads and updates the
tile index file from the Google Cloud. Users can choose if they want
to use a SQLite database which is going to be created in the local
installation path of the package or if they want to use the raw
csv file. The latter consumes more disk space (about 5 GB) but
speeds up future queries significantly, while the former
consumes less disk space (about 2 GB) but increases each query time.
The function additionally checks if the file is older than a month
and asks the user if they wish to update. Additionally, the
user can force an update by setting the corresponding option.
}
\author{
Darius Görgen, Dr. Fabian Löw (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
\keyword{internal}
